♻️ Started work on cleaning up types
