+++ libraries/tokens/src/types/index.ts
@@ -1,5 +1,7 @@
-import { typography } from '../base/typography'
-import { elevation } from '../base/elevation'
+import { typography } from '../../base/typography'
+import { elevation } from '../../base/elevation'
+
+export * from './borders'
 
 export type Typography = {
   color: string
@@ -86,29 +88,3 @@ export type Clickbound = {
 }
 
 export type Background = string
-
-export type Border = {
-  type?: 'border'
-  radius?: string | number
-  color?: string
-  width?: string | number
-  style?: 'solid'
-}
-
-export type Outline = {
-  type: 'outline'
-  color: string
-  width: string
-  style: 'dashed'
-  offset?: string
-}
-
-export type Bordergroup = {
-  type?: 'bordergroup'
-  top?: Omit<Border, 'radius'>
-  left?: Border
-  right?: Border
-  bottom?: Omit<Border, 'radius'>
-}
-
-export type Borders = Border | Outline | Bordergroup
